refactor(permissions): move Role/Grant provisioning into services.py
Per docs/styleguides/python.md (HackSoft service/selector pattern), business
logic that writes to the database belongs in services.py, not a seed module.
Move sync_roles, backfill_shelter_grants and backfill_global_role_members from
accounts/seed.py into accounts/services.py (which already imports
sync_group_permissions from seed, so the dependency direction is unchanged).
The post_migrate hook, manage.py sync_roles, and the tests now import from
accounts.services. ADR §2.9 updated accordingly.
