+++ apps/betterangels-backend/accounts/apps.py
@@ -3,12 +3,8 @@ from django.db.models.signals import post_migrate
 
 
 def _seed_on_migrate(sender: AppConfig, **kwargs: object) -> None:
-    from accounts.seed import (
-        backfill_global_role_members,
-        backfill_shelter_grants,
-        seed_permission_templates,
-        sync_roles,
-    )
+    from accounts.seed import seed_permission_templates
+    from accounts.services import backfill_global_role_members, backfill_shelter_grants, sync_roles
 
     seed_permission_templates()
     sync_roles()
