+++ apps/betterangels-backend/accounts/seed.py
@@ -125,65 +125,3 @@ def sync_group_permissions(*, organization: Organization | None = None) -> None:
             if {p.pk for p in permission_group.permissions.all()} != wanted:
                 permission_group.permissions.set(wanted)
                 logger.info("Synced permissions for group %s (%d perms)", permission_group.name, len(wanted))
-
-
-def sync_roles() -> None:
-    """Create or refresh the code-owned ``Role`` rows (ADR 0001 §2.2).
-
-    One row per :class:`~common.permissions.config.RoleDef` — global roles are
-    provisioned once, never per organization.  Idempotent: get_or_create each
-    ``Role``, then reconcile ``permissions`` and ``is_global`` from the RoleDef.
-    """
-    from accounts.models import Role
-    from shelters.groups import ROLES
-
-    with transaction.atomic():
-        for role_def in ROLES:
-            role, created = Role.objects.get_or_create(name=role_def.name)
-            wanted = set(_resolve_permissions(role_def.permissions))
-            perms_changed = {p.pk for p in role.permissions.all()} != wanted
-            global_changed = role.is_global != role_def.is_global
-            if perms_changed:
-                role.permissions.set(wanted)
-            if global_changed:
-                role.is_global = role_def.is_global
-                role.save(update_fields=["is_global"])
-            if created or perms_changed or global_changed:
-                logger.info("Synced Role %s (%d perms, global=%s)", role.name, len(wanted), role.is_global)
-
-
-def backfill_shelter_grants() -> None:
-    """Backfill ``Grant`` rows from legacy Shelter Operator memberships.
-
-    One ``Grant(user, role=Shelter Operator, scope=org)`` per member of an org's
-    Shelter Operator ``PermissionGroup``.  Idempotent (``get_or_create``).  Only
-    the scoped shelter role is converted here — every other role keeps its
-    ``PermissionGroup`` until its domain cutover (ADR 0001 §4).
-    """
-    from accounts.models import Grant, PermissionGroup, Role
-    from shelters.groups import SHELTER_OPERATOR_ROLE
-
-    role = Role.objects.get(name=SHELTER_OPERATOR_ROLE.name)
-    groups = PermissionGroup.objects.filter(template__name=SHELTER_OPERATOR_ROLE.name)
-    for group in groups.prefetch_related("user_set"):
-        for user in group.user_set.all():
-            grant, created = Grant.objects.get_or_create(principal_user=user, role=role, scope_org=group.organization)
-            if created:
-                logger.info("Backfilled Grant %s", grant)
-
-
-def backfill_global_role_members() -> None:
-    """Move Global Shelter Operator members onto the global Role group.
-
-    The GSO ``PermissionGroup`` was pinned to one arbitrary org; its members now
-    belong on the global Role's group, which is the global tier (ADR 0001 §2.1).
-    Idempotent (``user.groups.add``).
-    """
-    from accounts.models import PermissionGroup, Role
-    from shelters.groups import GLOBAL_SHELTER_OPERATOR_ROLE
-
-    role = Role.objects.get(name=GLOBAL_SHELTER_OPERATOR_ROLE.name)
-    groups = PermissionGroup.objects.filter(template__name=GLOBAL_SHELTER_OPERATOR_ROLE.name)
-    for group in groups.prefetch_related("user_set"):
-        for user in group.user_set.all():
-            user.groups.add(role)
