+++ apps/betterangels-backend/accounts/services.py
@@ -6,6 +6,7 @@ Reference: https://github.com/HackSoftware/Django-Styleguide#services
 
 from __future__ import annotations
 
+import logging
 import uuid
 from typing import TYPE_CHECKING, Any
 
@@ -18,7 +19,6 @@ from organizations.models import Organization, OrganizationOwner, OrganizationUs
 
 from .emails import base_url_for
 from .groups import ORG_ADMIN
-from .seed import sync_group_permissions
 from .models import (
     OrganizationProfile,
     OrgTypeChoices,
@@ -27,9 +27,12 @@ from .models import (
 )
 from .models import User as UserModel
 from .role_manager import OrgRoleManager
+from .seed import _resolve_permissions, sync_group_permissions
 
 if TYPE_CHECKING:
     from .models import User
+
+logger = logging.getLogger(__name__)
 
 
 # ── User provisioning ────────────────────────────────────────────────
@@ -513,3 +516,68 @@ def create_organization_service(
     )
 
     return user, organization
+
+
+# ── Role + Grant provisioning (ADR 0001 §2.2, §4 phase 1) ───────────────
+
+
+def sync_roles() -> None:
+    """Create or refresh the code-owned ``Role`` rows (ADR 0001 §2.2).
+
+    One row per :class:`~common.permissions.config.RoleDef` — global roles are
+    provisioned once, never per organization.  Idempotent: get_or_create each
+    ``Role``, then reconcile ``permissions`` and ``is_global`` from the RoleDef.
+    """
+    from accounts.models import Role
+    from shelters.groups import ROLES
+
+    with transaction.atomic():
+        for role_def in ROLES:
+            role, created = Role.objects.get_or_create(name=role_def.name)
+            wanted = set(_resolve_permissions(role_def.permissions))
+            perms_changed = {p.pk for p in role.permissions.all()} != wanted
+            global_changed = role.is_global != role_def.is_global
+            if perms_changed:
+                role.permissions.set(wanted)
+            if global_changed:
+                role.is_global = role_def.is_global
+                role.save(update_fields=["is_global"])
+            if created or perms_changed or global_changed:
+                logger.info("Synced Role %s (%d perms, global=%s)", role.name, len(wanted), role.is_global)
+
+
+def backfill_shelter_grants() -> None:
+    """Backfill ``Grant`` rows from legacy Shelter Operator memberships.
+
+    One ``Grant(user, role=Shelter Operator, scope=org)`` per member of an org's
+    Shelter Operator ``PermissionGroup``.  Idempotent (``get_or_create``).  Only
+    the scoped shelter role is converted here — every other role keeps its
+    ``PermissionGroup`` until its domain cutover (ADR 0001 §4).
+    """
+    from accounts.models import Grant, PermissionGroup, Role
+    from shelters.groups import SHELTER_OPERATOR_ROLE
+
+    role = Role.objects.get(name=SHELTER_OPERATOR_ROLE.name)
+    groups = PermissionGroup.objects.filter(template__name=SHELTER_OPERATOR_ROLE.name)
+    for group in groups.prefetch_related("user_set"):
+        for user in group.user_set.all():
+            grant, created = Grant.objects.get_or_create(principal_user=user, role=role, scope_org=group.organization)
+            if created:
+                logger.info("Backfilled Grant %s", grant)
+
+
+def backfill_global_role_members() -> None:
+    """Move Global Shelter Operator members onto the global Role group.
+
+    The GSO ``PermissionGroup`` was pinned to one arbitrary org; its members now
+    belong on the global Role's group, which is the global tier (ADR 0001 §2.1).
+    Idempotent (``user.groups.add``).
+    """
+    from accounts.models import PermissionGroup, Role
+    from shelters.groups import GLOBAL_SHELTER_OPERATOR_ROLE
+
+    role = Role.objects.get(name=GLOBAL_SHELTER_OPERATOR_ROLE.name)
+    groups = PermissionGroup.objects.filter(template__name=GLOBAL_SHELTER_OPERATOR_ROLE.name)
+    for group in groups.prefetch_related("user_set"):
+        for user in group.user_set.all():
+            user.groups.add(role)
